move hfc optimization checks to GlobalForecastingModel
